+++ packages/dsw-tdk/CHANGELOG.md
@@ -10,7 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [4.0.0]
 
-Released for version consistency with other DSW tools.
+### Fixed
+
+- Fixed `unpackage` with default output directory
+- Fixed terminating watch mode
+- Fixed non-printable characters handling
 
 ## [3.28.0]