+++ packages/dsw-tdk/CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 
+## [4.0.0]
+
+### Fixed
+
+- Fixed `unpackage` with default output directory
+- Fixed terminating watch mode
+- Fixed non-printable characters handling
+
 ## [3.28.0]
 
 Released for version consistency with other DSW tools.
@@ -357,3 +365,4 @@ Initial DSW Template Development Kit (versioned as part of the [DSW platform](ht
 [3.27.0]: /../../tree/v3.27.0
 [3.27.1]: /../../tree/v3.27.1
 [3.28.0]: /../../tree/v3.28.0
+[4.0.0]: /../../tree/v4.0.0